+++ src/main/java/com/epam/aidial/cfg/service/config/transfer/VersionedSchemaLoader.java
@@ -42,20 +42,11 @@ public class VersionedSchemaLoader {
     }
 
     public JsonNode loadSchema(String version) {
-        version = resolveVersion(version);
-        return schemaCache.computeIfAbsent(version, this::loadSchemaFromFile);
-    }
-
-    public String getEffectiveVersion(String version) {
-        version = resolveVersion(version);
-        return findNearestAvailableVersion(version);
-    }
-
-    private String resolveVersion(String version) {
         validateVersionFormat(version);
-        return LATEST_VERSION.equals(version)
-                ? findLatestVersion()
-                : version;
+        if (LATEST_VERSION.equals(version)) {
+            version = findLatestVersion();
+        }
+        return schemaCache.computeIfAbsent(version, this::loadSchemaFromFile);
     }
 
     private JsonNode loadSchemaFromFile(String version) {
